# swagger.yml
openapi: 3.0.0
...
servers:
  - url: https://computemachines.com/api
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/computemachines/Inventory/2.0.0
info:
  description: "An API that provides operations to manage and track inventory items and corresponding financial information for managerial accounting,"
  version: "3.1.0"
  title: Inventory API
  contact:
    email: tparker@computemachines.com
...
paths:
...
  /uniqs:
    post:
      tags: [UNIQ]
      description: Create a new, fully-formed UNIQ unit
      parameters:
        - name: from_sku
          in: query
          description: form uniq from this sku
          schema:
            type: string
            example: SKU00051
        - name: from_bin
          in: query
          description: form uniq from sku in this bin. Leave formed uniq in this bin. Bin should have one fewer sku.
          schema:
            type: string
            example: BIN00081
      requestBody:
        description: Unit to post to collection.
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                sku_id:
                  type: string
          application/json:
            schema:
              $ref: '#/components/schemas/UNIQ'
      responses:
        201:
          description: Successfully created a new unit
          headers:
            Location:
              description: Location of the created unit
              schema:
                type: string
                format: uri
                example: "https://computemachines.com/api/unit/UNIQ00012"
        404:
          description: Bin or parent sku does not exist.
          headers:
            Location:
              description: Location of the resource that needs to be created.
              schema:
                type: string
                format: uri
                example: "https://computemachines.com/api/bin/BIN0000013"
        409:
          description: Unit already exists
          headers:
            Location:
              description: Location of the existing unit
              schema:
                type: string
                format: uri
                example: "https://computemachines.com/api/unit/UNIQ00012"
...
  /uniq/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: UNIQ00081
    get:
      tags: [UNIQ]
      description: Get info on unit with this UNIQ code.
      responses:
        200:
          description: Successful pull of uniq info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
        404:
          description: The unit can not be found
    put:
      tags: [UNIQ]
      description: Update uniq info
      requestBody:
        content:
          application/json: 
            schema:
              $ref: '#/components/schemas/UNIQ'
      responses:
        200:
          description: Updated
        400:
          description: Bad format. Don't try to update uniq locations with this endpoint.
        404:
          description: uniq does not exist
    delete:
      description: Dissolve uniq back to sku or just delete unit
      parameters:
        - name: dissolve
          in: query
          description: Convert the UNIQ to SKU. If false then just delete UNIQ.
          required: true
          schema:
            type: boolean
      tags: [UNIQ]
      responses:
        200:
          description: Unit was deleted.
        404:
          description: The unit does not exist so can not be deleted.
...
components:
  schemas:
    BinProperties:
      description: "Additional properties"
      additionalProperties: true
      example:
        bin dimensions: ['100mm', '200mm', '60mm']
        process: Receive
    Bin:
      description: Complete information about a Bin 
      properties:
        id:
          type: string
          example: 'BIN0000012'
        contents:
          type: array
          items:
            oneOf:
              - type: object
                properties:
                  sku_id:
                    type: string
                    example: SKU000012
                  quantity:
                    type: integer
                    minimum: 1
                    example: 1
              - type: object
                properties:
                  batch_id:
                    type: string
                    example: BAT000012
                  quantity:
                    type: integer
                    minimum: 1
                    example: 1
              - type: object
                properties:
                  uniq_id:
                    type: string
                    example: SKU000012
                  quantity:
                    type: integer
                    enum: [1]
        props:
          $ref: '#/components/schemas/BinProperties'
    UnitProperties:
      description: 'Additional Properties'
      additionalProperties: true
      example:
        dc gain: 100db
...
    SKU:
      description: A definition of a physical unit.
      properties:
        unit type:
          type: string
          enum: [SKU]
        id:
          type: string
          example: SKU000081
        owned codes:
          type: array
          items:
            type: string
            description: Any code that maps to only this SKU. Can be a product number, ASIN, GTIN, IAN etc, as long as they do not reference any other SKUs. List must start with the SKU id.
          example: ["SKU000081", "12345678"]
        associated codes:
          type: array
          items:
            type: string
            description: Codes that should not be owned codes of any other SKU because they might be physically attached to this unit. Scan every barcode on the unit. A code can be associated with multiple SKU.
        name:
          type: string
          description: A useful, human-readable name for this item. 
          example: "Op-Amp (Through-Hole) - LM358"
        average unit original cost:
          type: string
          format: currency
          description: Asset cost per unit averaged over all units with this SKU at the time of each unit's production or purchase.
        props:
          $ref: '#/components/schemas/UnitProperties'
      required:
        - id
        - owned codes
        - name
    BatchProperties:
      properties:
        original cost:
          type: string
          format: currency
          description: Total asset cost at the time of production, purchase or assembly. Should be equal to the the sum of 'value consumed' in assembly or the sku original cost.
        original cost per unit:
          type: string
          format: currency
          description: original cost divided by the unit count.
        asset value:
          type: string
          format: currency
          description: 'Total asset value of this batch. Used for computing Assets and Cost Of Goods Sold for accounting. Must include effects of consumption and depreciation.'
        asset value per unit:
          type: string
          format: currency
          description: 'Asset value divided by unit count'
        unit count:
          type: integer
        assembly:
          description: "A set of units that contributed to the production of this batch. Assembled units must reduce their 'asset value' by the 'value consumed' of this unit"
          type: array
          items:
            type: object
            properties:
              unit:
                $ref: '#/components/schemas/Unit'
              value consumed:
                type: string
                format: currency
        assembly date:
          type: string
          format: date
        expiration date:
          type: string
          format: date
    Batch:
      description: A set of units of an SKU that for some reason need to be distinguised from other units with the same SKU. A unit can be in at most one batch. All units in a batch must not be in the same bin as any units of the same SKU that are not in the batch. Batches can be split, formed and dissolved.
      properties:
        unit type:
          type: string
          enum: [Batch]
        id:
          type: string
          example: BATCH00081
        bins:
          type: array
          items:
            type: object
            properties:
              bin:
                type: string
                example: BIN000076
              count:
                type: integer
                example: 10
        sku:
          $ref: '#/components/schemas/SKU'
        props:
          $ref: '#/components/schemas/BatchProperties'
      required:
        - sku
        - id
    UNIQ:
      description: "A single unit labeled with a globally uniq serial number prefixed by 'UNIQ'. Can optionally inherit properties from an SKU. A UNIQ label should be persistent across the lifetime of the part. A single item batch."
      properties:
        unit type:
          type: string
          enum: [UNIQ]
        id:
          type: string
          example: UNIQ00008
        bin:
          $ref: '#/components/schemas/Bin'
        sku:
          $ref: '#/components/schemas/SKU'
        name:
          type: string
          description: A useful, human-readable name for this item. 
          example: "Op-Amp (Through-Hole) - LM358"
        original cost:
          type: string
          format: currency
          description: Asset price at the time of production or purchase. Should be equal to the the sum of 'value consumed' in assembly or the sku original cost.
        asset value:
          type: string
          format: currency
          description: 'Used for computing Assets and Cost Of Goods Sold for accounting. Must include effects of consumption and depreciation.'
        assembly:
          description: "A set of units that contributed to the production of this unit. Assembled units must in total reduce their 'asset value' by the 'value consumed'"
          type: array
          items:
            type: object
            properties:
              unit:
                $ref: '#/components/schemas/Unit'
              value consumed:
                type: string
                format: currency
        
        assembly date:
          type: string
          format: date
        expiration date:
          type: string
          format: date
        props:
          $ref: '#/components/schemas/UnitProperties'
    Unit:
      description: Algerbraic data type.
      oneOf:
        - $ref: '#/components/schemas/SKU'
        - $ref: '#/components/schemas/Batch'
        - $ref: '#/components/schemas/UNIQ'
      discriminator:
        propertyName: unit type